Fix Vector's handling of Range parameters
Vector's handling of Range parameters was not compatible with Ruby 1.9.
This commit fixes Vector's handling of Range parameters in a way that is
compatible with both Ruby 1.8 and Ruby 1.9.

Before this commit, the beginning and end of a Range were obtained by
calling rb_ivar_get on the Range instance using the ID for "begin" or
"end".  It is amazing that this worked at all since these are not
instance variables, but rather methods.  They should be called
via rb_funcall3 instead of gotten ("getted") via rb_ivar_get.

git-svn-id: http://rb-gsl.rubyforge.org/svn/trunk/rb-gsl@118 6e764f74-f39f-46f8-8c54-8307d62afe8d